requirements.txt
alabaster==0.7.12
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
coverage==5.5
docutils==0.16
Flask==2.0.1
gitdb==4.0.7
GitPython==0.3.6
idna==3.2
imagesize==1.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
nose==1.3.7
numpy==1.21.0
packaging==21.0
pandas==1.3.0
Pillow==8.3.1
pydocstyle==6.1.1
Pygments==2.9.0
pyparsing==2.4.7
pypi-publisher==0.0.4
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
requests==2.26.0
six==1.16.0
smmap==4.0.0
snowballstemmer==2.1.0
Sphinx==4.1.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==1.26.6
Werkzeug==2.0.1
